Add older processing scripts to clarify provenance of PAM50 atlas files (#47)
* Copy scripts from dev/atlas/create_atlas (1284a9b)

* `create_atlas.m`: Add detailed info to dependencies section

This should be everything that's needed to run the script. :)

* `create_atlas.m`: Improve logging for iterative steps

Previously it was too difficult to tell what part of the script you were in (since it takes a long time, it's important to know what steps we're on).

* `README.md`: Add links to archival scripts used for atlas generation

Making it more discoverable -- though this doesn't fully document the exact procedures used. Instead, the
CHANGES.md file should be used for this purpose.